Update 09-re-render.html
Hey @kentcdodds , 

Thanks so much for building this course. I really appreciate it. I noticed your code throws the following error when the props to input is not set to readOnly or has an onChange handler. I went ahead and put it in. 

Would you approve this change? 

Thanks
Ajin